blends siminfo tests into database tests
